[Fixed] Various bugs with Object Factory, including BString's being converted to BSTRType_ANSI. And fixing some stored procedure calls to properly use SELECT with sf_ and CALL sp_

+++ src/ZoneServer/ObjectFactory.h
@@ -97,7 +97,7 @@ public:
     void					requestNewDefaultItem(ObjectFactoryCallback* ofCallback,uint32 familyId,uint32 typeId,uint64 parentId,uint16 planetId, const glm::vec3& position, const BString& customName = "");
     void					requestNewDefaultManufactureSchematic(ObjectFactoryCallback* ofCallback,uint32 schemCrc,uint64 parentId);
     void					requestNewWaypoint(ObjectFactoryCallback* ofCallback,BString name, const glm::vec3& coords,uint16 planetId,uint64 ownerId,uint8 wpType);
-    void					requestUpdatedWaypoint(ObjectFactoryCallback* ofCallback,uint64 wpId,BString name, const glm::vec3& coords,uint16 planetId,uint64 ownerId,uint8 activeStatus);
+    void					requestUpdatedWaypoint(ObjectFactoryCallback* ofCallback,uint64_t wpId,BString name, const glm::vec3& coords,uint16_t planetId,uint64_t ownerId,uint16_t activeStatus);
     void					requestNewTravelTicket(ObjectFactoryCallback* ofCallback,TicketProperties ticketProperties,uint64 parentId,uint16 planetId);
     void					requestNewResourceContainer(ObjectFactoryCallback* ofCallback,uint64 resourceId,uint64 parentId,uint16 planetId,uint32 amount);
     void					requestnewHarvesterbyDeed(ObjectFactoryCallback* ofCallback,Deed* deed,DispatchClient* client, float x, float y, float z, float dir, BString customName, PlayerObject* player);
